+++ tests/classes/Matrix.test.ts
@@ -7,8 +7,8 @@ before("initialize", async function() {
 });
 
 describe('matrix_constructor', function() {
-  it('float', function() {
-    const mat: Matrix<number> = new Matrix<number>({ n_cols: 2, n_rows: 2 }, "float");
+  it('double', function() {
+    const mat: Matrix<number> = new Matrix<number>({ n_cols: 2, n_rows: 2 }, "double");
   });
 
   it('int', function() {
@@ -17,8 +17,8 @@ describe('matrix_constructor', function() {
 });
 
 describe('matrix_init', function() {
-  it('float', async function() {
-    const mat: Matrix<number> = new Matrix<number>({ n_cols: 2, n_rows: 2 }, "float");
+  it('double', async function() {
+    const mat: Matrix<number> = new Matrix<number>({ n_cols: 2, n_rows: 2 }, "double");
     await mat.init();
   });
 
@@ -31,7 +31,7 @@ describe('matrix_init', function() {
 describe('matrix_methods', function() {
   it('fromArray', async function() {
     const arr: Array<number> = [1, 2, 3, 4, 5, 6];
-    const mat: Matrix<number> = new Matrix<number>({ n_cols: 3, n_rows: 2 }, "float");
+    const mat: Matrix<number> = new Matrix<number>({ n_cols: 3, n_rows: 2 }, "double");
     await mat.init();
     mat.fromArray(arr);
     ObjEqualArray(mat, arr);
@@ -39,7 +39,7 @@ describe('matrix_methods', function() {
 
   it('zeros', async function() {
     const arr: Array<number> = new Array(25).fill(0);
-    const mat: Matrix<number> = new Matrix<number>({ n_cols: 5, n_rows: 5 }, "float");
+    const mat: Matrix<number> = new Matrix<number>({ n_cols: 5, n_rows: 5 }, "double");
     await mat.init();
     mat.zeros();
     ObjEqualArray(mat, arr);
@@ -47,7 +47,7 @@ describe('matrix_methods', function() {
 
   it('ones', async function() {
     const arr: Array<number> = new Array(25).fill(1);
-    const mat: Matrix<number> = new Matrix<number>({ n_cols: 5, n_rows: 5 }, "float");
+    const mat: Matrix<number> = new Matrix<number>({ n_cols: 5, n_rows: 5 }, "double");
     await mat.init();
     mat.ones();
     ObjEqualArray(mat, arr);
@@ -59,7 +59,7 @@ describe('matrix_methods', function() {
                                 0, 0, 1, 0, 0,
                                 0, 0, 0, 1, 0,
                                 0, 0, 0, 0, 1];
-    const mat: Matrix<number> = new Matrix<number>({ n_cols: 5, n_rows: 5 }, "float");
+    const mat: Matrix<number> = new Matrix<number>({ n_cols: 5, n_rows: 5 }, "double");
     await mat.init();
     mat.eye();
     ObjEqualArray(mat, arr);
@@ -67,7 +67,7 @@ describe('matrix_methods', function() {
 
   it('fill', async function() {
     const arr: Array<number> = new Array(25).fill(21);
-    const mat: Matrix<number> = new Matrix<number>({ n_cols: 5, n_rows: 5 }, "float");
+    const mat: Matrix<number> = new Matrix<number>({ n_cols: 5, n_rows: 5 }, "double");
     await mat.init();
     mat.fill(21);
     ObjEqualArray(mat, arr);
@@ -79,7 +79,7 @@ describe('matrix_methods', function() {
                                 2, 3, 4, 5, 6,
                                 3, 4, 5, 6, 7,
                                 4, 5, 6, 7, 8];
-    const mat: Matrix<number> = new Matrix<number>({ n_cols: 5, n_rows: 5}, "float");
+    const mat: Matrix<number> = new Matrix<number>({ n_cols: 5, n_rows: 5}, "double");
     await mat.init();
     mat.funcFill((i, j) => i + j);
     ObjEqualArray(mat, arr);
@@ -98,8 +98,8 @@ describe('matrix_methods', function() {
                                          4, 2, 1, 7, 7,
                                          5, 3, 9, 9, 9];
 
-    const matA: Matrix<number> = new Matrix<number>({ n_cols: 5, n_rows: 5}, "float");
-    const matB: Matrix<number> = new Matrix<number>({ n_cols: 5, n_rows: 5}, "float");
+    const matA: Matrix<number> = new Matrix<number>({ n_cols: 5, n_rows: 5}, "double");
+    const matB: Matrix<number> = new Matrix<number>({ n_cols: 5, n_rows: 5}, "double");
     await matA.init();
     await matB.init();
